backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

# Vector DB
qdrant-client==1.11.0

# HTTP
httpx==0.27.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 문서 처리
PyMuPDF==1.24.10
python-docx==1.1.2

# 한국어 BM25 토크나이징
kiwipiepy==0.18.0
